# packages/contracts/openapi/child-safety-v1.yaml
openapi: 3.1.0
info:
  title: Child Safety API
  version: 1.0.0
paths:
  /v1/locations:resolve:
    get:
      summary: Resolve a city/state to geographic coordinates
      operationId: resolveLocation
      parameters:
        - $ref: '#/components/parameters/City'
        - $ref: '#/components/parameters/State'
      responses:
        '200':
          description: Resolved location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolvedLocation'
  /v1/heatmaps/overlay:
    get:
      summary: Get safety heatmap data for a specific area
      operationId: getHeatmapOverlay
      parameters:
        - $ref: '#/components/parameters/City'
        - $ref: '#/components/parameters/State'
        - name: radius_km
          in: query
          required: false
          schema: { type: number, minimum: 0.1, default: 8 }
        - name: crime_type
          in: query
          required: false
          schema: { type: string, default: all }
      responses:
        '200':
          description: Heatmap overlay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeatmapOverlay'

  # --- Identity Vault Endpoints ---
  /v1/vault/profiles:
    get:
      summary: List all child profiles in the vault
      operationId: listProfiles
      responses:
        '200':
          description: A list of child profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChildProfile'
    post:
      summary: Create a new child profile in the vault
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChildProfile'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildProfile'
...
  /v1/tracker/policies:
    post:
      summary: Save tracker policy settings for a dependent profile
      operationId: setTrackerPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackerPolicyRequest'
      responses:
        '200':
          description: Tracker policy persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerPolicyResponse'
  /v1/tracker/policies/{profileId}:
    get:
      summary: Get tracker policy for a dependent profile
      operationId: getTrackerPolicy
      parameters:
        - name: profileId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Tracker policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerPolicyResponse'
  /v1/tracker/history/{profileId}:
    get:
      summary: Get tracker location history with viewer-based privacy redaction
      operationId: getTrackerHistory
      parameters:
        - name: profileId
          in: path
          required: true
          schema: { type: string }
        - name: viewerRole
          in: query
          required: false
          schema: { type: string, default: guardian }
      responses:
        '200':
          description: Tracker history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerHistoryResponse'
  /v1/tracker/history/{profileId}:export:
    post:
      summary: Create signed tracker history export for authority or guardian workflows
      operationId: exportTrackerHistory
      parameters:
        - name: profileId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackerExportRequest'
      responses:
        '200':
          description: Signed export generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerExportResponse'
  /v1/tracker/exports/{exportId}:
    get:
      summary: Retrieve signed tracker export payload by export ID with expiry enforcement
      operationId: getTrackerExport
      parameters:
        - name: exportId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Signed export bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerExportBundleResponse'
        '404':
          description: Export not found
        '410':
          description: Export expired
  /v1/tracker/audit/{profileId}:
    get:
      summary: Get tracker access and export audit events
      operationId: getTrackerAudit
      parameters:
        - name: profileId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Tracker audit entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerAuditResponse'

  # --- Secure Communication Endpoints ---
...
  /v1/alerts/help:
    post:
      summary: Create a discreet emergency help alert
      operationId: createHelpAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelpAlertRequest'
      responses:
        '201':
          description: Help alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelpAlertResponse'
  /v1/alerts/help/{alertId}/locations:
    post:
      summary: Append a location sample for an active help alert
      operationId: appendHelpAlertLocation
      parameters:
        - name: alertId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelpAlertLocationRequest'
      responses:
        '202':
          description: Location accepted for alert tracking
...
components:
  parameters:
    City:
      name: city
      in: query
      schema: { type: string, default: Columbus }
    State:
      name: state
      in: query
      schema: { type: string, default: OH }
  schemas:
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number }
        lng: { type: number }
    ResolvedLocation:
      type: object
      required: [city, state, location]
      properties:
        city: { type: string }
        state: { type: string }
        location: { $ref: '#/components/schemas/Location' }
    HeatmapCell:
      type: object
      required: [lat, lng, intensity, count, crime_types, properties]
      properties:
        lat: { type: number }
        lng: { type: number }
        intensity: { type: number, minimum: 0, maximum: 1 }
        count: { type: integer, minimum: 0 }
        crime_types:
          type: array
          items: { type: string }
        properties: { type: object, additionalProperties: true }
    HeatmapOverlay:
      type: object
      required: [location, radius_km, filters, cell_count, cells]
      properties:
        location: { $ref: '#/components/schemas/Location' }
        radius_km: { type: number }
        filters: { type: object, additionalProperties: true }
        cell_count: { type: integer }
        cells:
          type: array
          items: { $ref: '#/components/schemas/HeatmapCell' }

    # --- New Schemas for Consistently Platform-Agnostic Models ---
    Attachment:
      type: object
      required: [id, name, content_type]
      properties:
        id: { type: string }
        name: { type: string }
        content_type: { type: string }
        url: { type: string, format: uri }
    ChildProfile:
      type: object
      required: [name, dob]
      properties:
        id: { type: string }
        name: { type: string }
        dob: { type: string }
        description: { type: string }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/Attachment' }
        metadata: { type: object, additionalProperties: true }
    LocationEvent:
      type: object
      required: [profileId, location, timestamp]
      properties:
        profileId: { type: string }
        location: { $ref: '#/components/schemas/Location' }
        timestamp: { type: string, format: date-time }
        activity_type: { type: string, enum: [stationary, walking, running, driving, unknown] }
        battery_level: { type: number, minimum: 0, maximum: 100 }
...
    HelpAlertLocation:
      type: object
      required: [lat, lng, captured_at]
      properties:
        lat: { type: number }
        lng: { type: number }
        accuracy_m: { type: number, minimum: 0 }
        captured_at: { type: string, format: date-time }
    HelpAlertRequest:
      type: object
      required: [profileId, recipientId, encryptedPayload]
      properties:
        profileId: { type: string }
        recipientId: { type: string }
        encryptedPayload: { type: string, description: "Ciphertext payload for the emergency message" }
        channel: { type: string, default: guardian }
        hiddenMode: { type: boolean, default: true }
        retentionMinutes: { type: integer, minimum: 5, maximum: 10080, description: "Retention duration in minutes" }
        location:
          $ref: '#/components/schemas/HelpAlertLocation'
    HelpAlertResponse:
      type: object
      required: [alertId, status, createdAt, expiresAt]
      properties:
        alertId: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    HelpAlertLocationRequest:
      type: object
      required: [location]
      properties:
        location:
          $ref: '#/components/schemas/HelpAlertLocation'
...
    TrackerPolicyRequest:
      type: object
      required: [profileId]
      properties:
        profileId: { type: string }
        mode: { type: string, enum: [off, emergency-only, scheduled, continuous-guardian], default: emergency-only }
        retentionHours: { type: integer, minimum: 1, maximum: 720, default: 24 }
        precisionMode: { type: string, enum: [high-emergency, balanced, coarse-default], default: high-emergency }
        allowAuthorityExport: { type: boolean, default: true }
        allowedViewers:
          type: array
          items: { type: string }
    TrackerPolicyResponse:
      type: object
      required: [profileId, mode, retentionHours, precisionMode, allowAuthorityExport, allowedViewers, updatedAt]
      properties:
        profileId: { type: string }
        mode: { type: string }
        retentionHours: { type: integer }
        precisionMode: { type: string }
        allowAuthorityExport: { type: boolean }
        allowedViewers:
          type: array
          items: { type: string }
        updatedAt: { type: string, format: date-time }
    TrackerHistoryResponse:
      type: object
      required: [profileId, viewerRole, eventCount, events]
      properties:
        profileId: { type: string }
        viewerRole: { type: string }
        eventCount: { type: integer, minimum: 0 }
        events:
          type: array
          items:
            $ref: '#/components/schemas/LocationEvent'
    TrackerExportRequest:
      type: object
      required: [requestedBy]
      properties:
        requestedBy: { type: string }
        role: { type: string, default: authority }
        scope: { type: string, default: last-24h }
        expiresInMinutes: { type: integer, minimum: 5, maximum: 1440, default: 60 }
    TrackerExportResponse:
      type: object
      required: [exportId, profileId, requestedBy, scope, generatedAt, expiresAt, eventCount, checksum, signature]
      properties:
        exportId: { type: string }
        profileId: { type: string }
        requestedBy: { type: string }
        scope: { type: string }
        generatedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        eventCount: { type: integer, minimum: 0 }
        checksum: { type: string }
        signature: { type: string }
    TrackerExportBundleResponse:
      type: object
      required: [exportId, profileId, requestedBy, role, scope, generatedAt, expiresAt, eventCount, checksum, signature, signatureValid, events]
      properties:
        exportId: { type: string }
        profileId: { type: string }
        requestedBy: { type: string }
        role: { type: string }
        scope: { type: string }
        generatedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        eventCount: { type: integer, minimum: 0 }
        checksum: { type: string }
        signature: { type: string }
        signatureValid: { type: boolean }
        events:
          type: array
          items:
            type: object
            additionalProperties: true
    TrackerAuditEntry:
      type: object
      required: [id, profileId, action, actor, createdAt, metadata]
      properties:
        id: { type: string }
        profileId: { type: string }
        action: { type: string }
        actor: { type: string }
        createdAt: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: true }
    TrackerAuditResponse:
      type: object
      required: [profileId, count, entries]
      properties:
        profileId: { type: string }
        count: { type: integer, minimum: 0 }
        entries:
          type: array
          items:
            $ref: '#/components/schemas/TrackerAuditEntry'